refactored nutrition calculator to typescript
