+++ src/routes/bmr/+page.server.ts
@@ -1,10 +1,10 @@
 import { fail, type Actions } from '@sveltejs/kit';
-import type { Action, PageServerLoad } from './$types';
 import { clearTodos, getTodos, removeTodo } from '$lib/server/database';
 import { addTodo } from '../../lib/server/database';
-import NutritionCalculator from '../../utils/NutritionCalculatorTypescript';
+import { calculateMacronutrients } from '../../utils/NutritionCalculations';
+('../../utils/NutritionCalculations');
 
-export const load: PageServerLoad = async () => {
+export const load: any = async () => {
 	const todos = getTodos();
 	return { todos };
 };
@@ -13,7 +13,6 @@ export const actions: Actions = {
 	addTodo: async ({ request }) => {
 		const formData = await request.formData();
 
-		// (bmrType = 'calculateBMRWithBF1'),
 		// 	(weight = 145),
 		// 	(height = 64),
 		// 	(age = 29),
@@ -21,7 +20,6 @@ export const actions: Actions = {
 		// 	(bf = 0.12),
 		// 	(activityLevel = 'sedentary');
 
-		const bmrType = String(formData.get('bmrType'));
 		const weight = Number(formData.get('weight'));
 		const height = Number(formData.get('height'));
 		const age = Number(formData.get('age'));
@@ -29,11 +27,10 @@ export const actions: Actions = {
 		const bf = Number(formData.get('bf'));
 		const activityLevel = String(formData.get('activityLevel'));
 
-		if (!bmrType || !weight || !height || !age || !sex || !bf || !activityLevel)
+		if (!weight || !height || !age || !sex || !bf || !activityLevel)
 			return fail(400, { weight, missing: true });
-		const calculator = new NutritionCalculator();
 		const results: string = JSON.stringify(
-			calculator.calculateMacronutrients(bmrType, weight, height, age, sex, bf, activityLevel)
+			calculateMacronutrients(weight, height, age, sex, bf, activityLevel)
 		);
 		addTodo(results);
